added a electrical_utilities.h and added some probability functions

+++ lab 5 Genetic Algorithms/Genetic Algorithms/Genetic Algorithms/include/Utilities.h
@@ -135,7 +135,7 @@ void print_dif_amount(vector<T>& a, vector<T>& b)
 
 std::map<int, PipCount>* pip_probability_counter(int sampleSize, int numDice, int numFaces, bool outputData = false);
 
-float bayesian_probability(float sensitivity, float specificity, float testPositive);
+double bayesian_probability(double sensitivity, double specificity, double testPositive);
 
 ///Get the number of matches between a and b
 template<typename T>
